requirements.txt
selenium
pytest
pyautogui
pytesseract
Pillow
opencv-python
pyperclip
requests>=2.31.0
aiohttp>=3.9.0
pytest-asyncio>=0.23.0